// src/types/types.ts
export interface OffsetStorage {
	connect(): Promise<void>
	disconnect(): Promise<void>
	commitOffset(commit: OffsetCommit): Promise<void>
	getOffset(
		groupId: string,
		topic: string,
		partition: number,
	): Promise<ResumeToken | null>
	getOffsets(groupId: string): Promise<OffsetCommit[]>
	isConnected?(): boolean
	getConnectionInfo?(): { database: string; collection: string }
}

export interface OffsetCommit {
	topic: string
	partition: number
	groupId: string
	offset: ResumeToken
	timestamp: Date
}